+++ DelvUI/Helpers/JobsHelper.cs
@@ -283,7 +283,7 @@ namespace DelvUI.Helpers
             return 0;
         }
 
-        public static uint RoleIconIDForBattleCompanion => 62041;
+        public static uint RoleIconIDForBattleCompanion => 62043;
 
         public static Dictionary<uint, JobRoles> JobRolesMap = new Dictionary<uint, JobRoles>()
         {